Adding more tests.
git-svn-id: https://outlast.svn.beanstalkapp.com/mozajik-private/trunk@303 1177bc47-7a3e-47f4-a2ea-9a189b2012c5

+++ system/plugins/_global/library/cache.lib.php
@@ -9,43 +9,31 @@
 class zajlib_cache extends zajLibExtension{
 
 	/**
-	 * This method will remove all object cache files. If the $class parameter is specified, the only the cache for that class will be cleared.
-	 * @param string $class An optional CaseSensitive parameter which specifies the name of the class to be cleared.
-	 * @return integer Returns the number of files cleared.
+	 * This method will remove all object cache files. If the $class parameter is specified, only the cache for that class will be cleared.
+	 * @param bool|string $class An optional CaseSensitive parameter which specifies the name of the class to be cleared.
+	 * @return bool Will alway return true.
 	 */
 	public function clear_objects($class=false){
 			// check to see if $class is valid
-				if($class && (!class_exists($class) || strstr($class, '.') !== false || strstr($class, '/') !== false)) return $this->zajlib->error("Invalid class name given while trying to clear the cache!");
+				if($class && (!class_exists($class) || strstr($class, '.') !== false || strstr($class, '/') !== false)) return $this->zajlib->warning("Invalid class name given while trying to clear the cache!");
 			// load file handler
-				//$this->zajlib->load->library('file');
-				//$all_files = $this->zajlib->file->get_files_in_dir($this->zajlib->basepath."cache/object/".$class."/", true);
 				system("rm -R ".$this->zajlib->basepath."cache/object/".$class."/");
-			// delete all files in cache folder
-				/**$count = 0;
-				foreach($all_files as $file){
-					$fdata = pathinfo($file);
-					if($fdata['extension'] == "cache"){
-						unlink($file);
-						$count++;
-					}
-				}**/
-		return $this->zajlib->basepath."cache/object/".$class."/";
+		return true;
 	}
 
 	/**
 	 * This method will remove a specific object cache file.
-	 * @param string $class The CaseSensitive parameter which specifies the name of the class to be cleared.
-	 * @param string $class The id parameter which specifies the specific id of the object to be cleared.
+	 * @param string $class_name The id parameter which specifies the specific id of the object to be cleared.
+	 * @param string $id The id parameter which specifies the specific id of the object to be cleared.
 	 * @return boolean Returns true if something was deleted, false otherwise.
 	 */
-	public function clear_object($class, $id){
+	public function clear_object($class_name, $id){
+		// check to see if $class is valid
+			if($class_name && (!class_exists($class_name) || strstr($class_name, '.') !== false || strstr($class_name, '/') !== false)) return $this->zajlib->warning("Invalid class name given while trying to clear an object from the cache!");
+			if(strstr($id, '.') !== false) return $this->zajlib->warning("Could not delete cache file. Invalid characters detected in file name.");
 		// Generate my path
 			$filename = $this->zajlib->file->get_id_path($this->zajlib->basepath."cache/object/".$class_name, $id.".cache", false, CACHE_DIR_LEVEL);
 		// Try to delete and return result
 			return @unlink($filename);
 	}
-	
-	
 }
-
-?>
